Document and improve the testDF function
The output will no longer show up in HTML tables, but will be shown in the Dodona diff. We still only use the `head` of the dataframes, but a message was added to explain it to the students. I also removed the `ignore_col/row_order` arguments, since TRUE is the default for those arguments in dplyr and the way it was implemented, they couldn't be passed to the custom comparator.