Fix libraries loaded by test code being available in the student environment
The way I did this (which is not trivial at all) is explained in the (new) comment at the top of context.R. One element which I'm not entirely happy with is that I had to duplicate the fix over both context functions. There were problems related to `preExec` which I was not able to fix however.