new contextWithParsing function added
This function takes a list of named codeblocks containing the testing function(s) per exercise section. To obtain these sections, the student code is being parsed (split on section titles following the pattern "### sectionname ###"). The section name has to match the named codeblock in the testcases list in order to run the testing functions.