Use custom method when comparing data frames and ignoring row order
Because small errors in floating point numbers can mess up the sort
then compare method, we need to be a little bit smarter. This method
sorts and then looks in the vicinity when comparing. This vicinity is
still implemented quite naively. We store the index before the first
not used row and check from there. There is no real stop condition, so
if the row is not found in the first few possibilities, we will still
check all remaining possibilities. However, this will almost certainly
stop the process, since it is quite unlikely this doesn't result in
the row not being matched.